Work

+++ EShop.Core/Catalog/Categories/Domain/ProductCategory.cs
@@ -9,7 +9,7 @@ namespace EShop.Core.Catalog.Categories.Domain
 
         public int CategoryId { get; set; }
 
-        public byte DisplayOrder { get; set; }
+        public int DisplayOrder { get; set; }
 
         public Product Product { get; set; }
